# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContaNova ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cuentas:
    get:
      operationId: listCuentas
      tags: [cuentas]
      summary: Listar plan de cuentas
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Plan de cuentas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CuentaContable"
    post:
      operationId: createCuenta
      tags: [cuentas]
      summary: Crear cuenta contable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CuentaContableInput"
      responses:
        "201":
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuentaContable"
  /cuentas/{id}:
    patch:
      operationId: updateCuenta
      tags: [cuentas]
      summary: Actualizar cuenta contable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CuentaContableUpdate"
      responses:
        "200":
          description: Cuenta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CuentaContable"
    delete:
      operationId: deleteCuenta
      tags: [cuentas]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ─── MOVIMIENTOS CONTABLES ────────────────────────────────────────────────────
  /movimientos:
    get:
      operationId: listMovimientos
      tags: [movimientos]
      summary: Listar comprobantes contables
      parameters:
        - name: tipo
          in: query
          schema:
            type: string
        - name: fechaDesde
          in: query
          schema:
            type: string
            format: date
        - name: fechaHasta
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movimiento"
    post:
      operationId: createMovimiento
      tags: [movimientos]
      summary: Crear comprobante contable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovimientoInput"
      responses:
        "201":
          description: Movimiento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movimiento"
  /movimientos/{id}:
    get:
      operationId: getMovimiento
      tags: [movimientos]
      summary: Obtener movimiento contable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movimiento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movimiento"
    delete:
      operationId: deleteMovimiento
      tags: [movimientos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado

  # ─── CRM ─────────────────────────────────────────────────────────────────────
  /crm/oportunidades:
    get:
      operationId: listOportunidades
      tags: [crm]
      summary: Listar oportunidades CRM
      parameters:
        - name: etapa
          in: query
          schema:
            type: string
            enum: [prospecto, calificado, propuesta, negociacion, ganado, perdido]
      responses:
        "200":
          description: Lista de oportunidades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Oportunidad"
    post:
      operationId: createOportunidad
      tags: [crm]
      summary: Crear oportunidad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OportunidadInput"
      responses:
        "201":
          description: Oportunidad creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oportunidad"
  /crm/oportunidades/{id}:
    patch:
      operationId: updateOportunidad
      tags: [crm]
      summary: Actualizar oportunidad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OportunidadUpdate"
      responses:
        "200":
          description: Oportunidad actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oportunidad"
    delete:
      operationId: deleteOportunidad
      tags: [crm]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Eliminado
...
components:
  schemas:
...
    CuentaContable:
      type: object
      required: [id, codigo, nombre, tipo, activo]
      properties:
        id:
          type: integer
        codigo:
          type: string
        nombre:
          type: string
        tipo:
          type: string
          enum: [activo, pasivo, patrimonio, ingreso, egreso]
        descripcion:
          type: ["string", "null"]
        cuentaPadreId:
          type: ["integer", "null"]
        activo:
          type: boolean
        saldo:
          type: ["number", "null"]
    CuentaContableInput:
      type: object
      required: [codigo, nombre, tipo]
      properties:
        codigo:
          type: string
          minLength: 1
        nombre:
          type: string
          minLength: 1
        tipo:
          type: string
          enum: [activo, pasivo, patrimonio, ingreso, egreso]
        descripcion:
          type: string
        cuentaPadreId:
          type: integer
    CuentaContableUpdate:
      type: object
      properties:
        codigo:
          type: string
        nombre:
          type: string
        tipo:
          type: string
          enum: [activo, pasivo, patrimonio, ingreso, egreso]
        descripcion:
          type: string
        activo:
          type: boolean

    # ─── MOVIMIENTOS CONTABLES ────────────────────────────────────────────────
    MovimientoLinea:
      type: object
      required: [cuentaId, cuentaCodigo, cuentaNombre, debito, credito]
      properties:
        cuentaId:
          type: integer
        cuentaCodigo:
          type: string
        cuentaNombre:
          type: string
        debito:
          type: number
        credito:
          type: number
        descripcion:
          type: ["string", "null"]
    MovimientoLineaInput:
      type: object
      required: [cuentaId, debito, credito]
      properties:
        cuentaId:
          type: integer
        debito:
          type: number
        credito:
          type: number
        descripcion:
          type: string
    Movimiento:
      type: object
      required: [id, numero, tipo, fecha, descripcion, totalDebito, totalCredito, createdAt]
      properties:
        id:
          type: integer
        numero:
          type: string
        tipo:
          type: string
        fecha:
          type: string
        descripcion:
          type: string
        totalDebito:
          type: number
        totalCredito:
          type: number
        lineas:
          type: array
          items:
            $ref: "#/components/schemas/MovimientoLinea"
        createdAt:
          type: string
    MovimientoInput:
      type: object
      required: [tipo, fecha, descripcion, lineas]
      properties:
        tipo:
          type: string
        fecha:
          type: string
          format: date
        descripcion:
          type: string
        lineas:
          type: array
          items:
            $ref: "#/components/schemas/MovimientoLineaInput"

    # ─── CRM ──────────────────────────────────────────────────────────────────
    Oportunidad:
      type: object
      required:
        - id
        - titulo
        - clienteId
        - clienteNombre
        - etapa
        - valor
        - probabilidad
        - fechaCierre
        - createdAt
      properties:
        id:
          type: integer
        titulo:
          type: string
        clienteId:
          type: integer
        clienteNombre:
          type: string
        etapa:
          type: string
          enum: [prospecto, calificado, propuesta, negociacion, ganado, perdido]
        valor:
          type: number
        probabilidad:
          type: integer
        fechaCierre:
          type: string
        descripcion:
          type: ["string", "null"]
        responsable:
          type: ["string", "null"]
        createdAt:
          type: string
    OportunidadInput:
      type: object
      required: [titulo, clienteId, etapa, valor, probabilidad, fechaCierre]
      properties:
        titulo:
          type: string
          minLength: 1
        clienteId:
          type: integer
        etapa:
          type: string
          enum: [prospecto, calificado, propuesta, negociacion, ganado, perdido]
        valor:
          type: number
        probabilidad:
          type: integer
        fechaCierre:
          type: string
          format: date
        descripcion:
          type: string
        responsable:
          type: string
    OportunidadUpdate:
      type: object
      properties:
        titulo:
          type: string
        etapa:
          type: string
          enum: [prospecto, calificado, propuesta, negociacion, ganado, perdido]
        valor:
          type: number
        probabilidad:
          type: integer
        fechaCierre:
          type: string
        descripcion:
          type: string
        responsable:
          type: string
...