# lib/api-zod/src/generated/types/cuentaContable.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContaNova ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { CuentaContableTipo } from './cuentaContableTipo';

export interface CuentaContable {
  id: number;
  codigo: string;
  nombre: string;
  tipo: CuentaContableTipo;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  cuentaPadreId?: number | null;
  activo: boolean;
  /** @nullable */
  saldo?: number | null;
}