# lib/api-zod/src/generated/types/cuentaContableInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContaNova ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { CuentaContableInputTipo } from './cuentaContableInputTipo';

export interface CuentaContableInput {
  /** @minLength 1 */
  codigo: string;
  /** @minLength 1 */
  nombre: string;
  tipo: CuentaContableInputTipo;
  descripcion?: string;
  cuentaPadreId?: number;
}