# lib/api-zod/src/generated/types/movimientoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContaNova ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { MovimientoLineaInput } from './movimientoLineaInput';

export interface MovimientoInput {
  tipo: string;
  fecha: Date;
  descripcion: string;
  lineas: MovimientoLineaInput[];
}