# lib/api-zod/src/generated/types/movimientoLineaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContaNova ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface MovimientoLineaInput {
  cuentaId: number;
  debito: number;
  credito: number;
  descripcion?: string;
}