# lib/api-zod/src/generated/types/oportunidadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ContaNova ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { OportunidadInputEtapa } from './oportunidadInputEtapa';

export interface OportunidadInput {
  /** @minLength 1 */
  titulo: string;
  clienteId: number;
  etapa: OportunidadInputEtapa;
  valor: number;
  probabilidad: number;
  fechaCierre: Date;
  descripcion?: string;
  responsable?: string;
}